refactor : enum 자료형 DB에서는 int 사용,  코드 상에서는 enum 사용으로 변경

+++ hotel/src/main/java/com/mjc/hotel/review/dto/ReviewPhotoDto.java
@@ -18,7 +18,7 @@ public class ReviewPhotoDto {
     private Long id;
     private Long reviewId;
     private String photoPath;
-    private Long photoOrder;
+    private Integer photoOrder;
 
     public static ReviewPhotoDto from(ReviewPhoto photo) {
         if (photo == null) {
